tests/python/.tungsten-build-emdmwik7/tungstenkit/requirements.txt
anyio==3.6.2 ; python_version >= "3.7" and python_version < "4.0"
appdirs==1.4.4 ; python_version >= "3.7" and python_version < "4.0"
attrs==22.2.0 ; python_version >= "3.7" and python_version < "4.0"
beautifulsoup4==4.12.2 ; python_version >= "3.7" and python_version < "4.0"
binaryornot==0.4.4 ; python_version >= "3.7" and python_version < "4.0"
cattrs==22.2.0 ; python_version >= "3.7" and python_version < "4.0"
certifi==2022.12.7 ; python_version >= "3.7" and python_version < "4.0"
chardet==5.1.0 ; python_version >= "3.7" and python_version < "4.0"
charset-normalizer==3.1.0 ; python_version >= "3.7" and python_version < "4.0"
click==8.1.3 ; python_version >= "3.7" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.7" and python_version < "4.0" and (sys_platform == "win32" or platform_system == "Windows")
dill==0.3.6 ; python_version >= "3.7" and python_version < "4.0"
docker==6.0.1 ; python_version >= "3.7" and python_version < "4.0"
exceptiongroup==1.1.1 ; python_version >= "3.7" and python_version < "3.11"
fastapi==0.95.1 ; python_version >= "3.7" and python_version < "4.0"
fasteners==0.18 ; python_version >= "3.7" and python_version < "4.0"
filelock==3.12.0 ; python_version >= "3.7" and python_version < "4.0"
furl==2.1.3 ; python_version >= "3.7" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.7" and python_version < "4.0"
idna==3.4 ; python_version >= "3.7" and python_version < "4.0"
importlib-metadata==6.6.0 ; python_version >= "3.7" and python_version < "3.10"
jinja2==3.1.2 ; python_version >= "3.7" and python_version < "4.0"
loguru==0.7.0 ; python_version >= "3.7" and python_version < "4.0"
markdown-it-py==2.2.0 ; python_version >= "3.7" and python_version < "4.0"
markdown==3.4.3 ; python_version >= "3.7" and python_version < "4.0"
markdownify==0.11.6 ; python_version >= "3.7" and python_version < "4.0"
markupsafe==2.1.2 ; python_version >= "3.7" and python_version < "4.0"
mdurl==0.1.2 ; python_version >= "3.7" and python_version < "4.0"
orderedmultidict==1.0.1 ; python_version >= "3.7" and python_version < "4.0"
packaging==23.1 ; python_version >= "3.7" and python_version < "4.0"
pathspec==0.10.3 ; python_version >= "3.7" and python_version < "4.0"
pillow==9.5.0 ; python_version >= "3.7" and python_version < "4.0"
pydantic==1.10.7 ; python_version >= "3.7" and python_version < "4.0"
pygments==2.15.1 ; python_version >= "3.7" and python_version < "4.0"
python-multipart==0.0.6 ; python_version >= "3.7" and python_version < "4.0"
pywin32==306 ; python_version >= "3.7" and python_version < "4.0" and sys_platform == "win32"
pyyaml==6.0 ; python_version >= "3.7" and python_version < "4.0"
requests-toolbelt==0.10.1 ; python_version >= "3.7" and python_version < "4.0"
requests==2.30.0 ; python_version >= "3.7" and python_version < "4.0"
rich==13.3.5 ; python_version >= "3.7" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.7" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.7" and python_version < "4.0"
soupsieve==2.4.1 ; python_version >= "3.7" and python_version < "4.0"
starlette==0.26.1 ; python_version >= "3.7" and python_version < "4.0"
tabulate==0.9.0 ; python_version >= "3.7" and python_version < "4.0"
typing-extensions==4.6.2 ; python_version >= "3.7" and python_version < "4.0"
urllib3==1.26.15 ; python_version >= "3.7" and python_version < "4.0"
uvicorn==0.22.0 ; python_version >= "3.7" and python_version < "4.0"
w3lib==2.1.1 ; python_version >= "3.7" and python_version < "4.0"
websocket-client==1.5.1 ; python_version >= "3.7" and python_version < "4.0"
win32-setctime==1.1.0 ; python_version >= "3.7" and python_version < "4.0" and sys_platform == "win32"
zipp==3.15.0 ; python_version >= "3.7" and python_version < "3.10"
